backend/requirements.txt
fastapi==0.95.0
mysql_connector_repackaged==0.3.1
pandas==1.4.4
pydantic==1.10.6
PyJWT==2.6.0
uvicorn==0.21.1
